+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed an error when comparing a Period to a timedelta in PyPy.
+
+
 ## [1.4.0] - 2018-01-22
 
 ### Changed